// api/generated/models/moduleResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lingostan API
 * OpenAPI spec version: 1.0
 */

export interface ModuleResponseDto {
  id: number;
  title: string;
  description: string;
  order: number;
  difficulty: string;
  imageUrl: string;
  estimatedDuration: number;
  totalLessons: number;
  totalExercises: number;
  isActive: boolean;
  language: number;
  progress: number;
  completed: boolean;
  lessons: number[];
  createdAt: string;
}